Refactor #18: app structure and add meal component
Completely overhauled the application structure to make way for the addition of food, meal, and pet pages. These changes also included the creation of a new basic meal component to enhance the UI. Login features and Vuetify dependencies were removed as these are no longer needed in the new structure.

Various style and format changes were also made to improve consistency and readability. This includes the addition of a universal color scheme within a new 'assets' directory and the introduction of a navigation bar component for easy navigation between pages.

The changes bring a much-needed modernization and optimization to the codebase, paving the way for future enhancements and simplifying maintenance.