Rename #18: NavBar to ToolBar; Add PageHeader, enhance MealInfo and pages
NavBar component is renamed to ToolBar for better semantics. A new component PageHeader is added which was utilized in meal-page.vue. MealInfo component's height has been reduced and border-radius now uses a variable from _colors.sass. The use of NavBar has been updated to ToolBar in all occurrences in pages.

Additional changes were made in meal-page.vue and _colors.sass, providing a more detailed and descriptive layout and design for the meal page. Repeated usage of the meal component offers a more dynamic look to the page for testing. The addition of new variables in _colors.sass made the design more interchangeable and maintainable.