feat #18: pet Color. various backend and frontend files to enhance API functionalities
A new model attribute 'color' was added to the 'Pet' model in the backend and the frontend was adjusted accordingly. User authentication was incorporated with necessary changes in the backend files. The new 'UserSerializer' is used to send user data to and from the backend.

Unneeded axios calls were removed from the frontend Vue components, replaced with fewer fetch calls instead. The meal-card Vue component was refactored to improve performance and maintainability.

Also, added necessary CORS configurations in the backend to handle resource sharing with the frontend.

Finally, multiple stylistic conventions were corrected for overall readability and consistency of the code. This includes changes like single quotes to double quotes in many python files, re-ordering import statements as per PEP8 guideline etc.

These changes are done to provide a smoother user experience, allowing for quicker updates and enhanced overall performance of the application. The changes also make it easier for other developers to read, understand, and update the codebase efficiently in the future.