Refactor component file structure and update navigation in tool-bar.vue
Refactored the folder structure of the components to ensure that all component files are in the same folder; unnecessary capitalization of folders is removed. This improves file finding efficiency and ensures a consistent naming convention. The diff also shows a shift from use of <button> to <Nuxtlink> in tool-bar.vue to facilitate user navigation to different pages upon clicking button-like text. Additionally, a new default.vue layout is introduced to utilize the slot concept, placing repetitively used code like page-header and tool-bar in one slot to maximize reusability and minimize redundancy.