Feat #18: Delete Meal, refactor API
This commit primarily includes the feature of deleting a meal from the meal card. In addition, it refactors usage of API URLs in the code.

The deletion functionality has been added by introducing a new method `deleteMeal` in `components/meal-card.vue`. It sends a DELETE request to the server when called. Once the deletion has been successful, it emits a 'refresh-list' event to signal the need for a UI refresh on the meals list. This event is captured in `pages/meal-page.vue` to refresh the meals list.

The refactor for the API URLs is done by creating a new file `helpers/api.js` which exports the base URL as a constant. This base URL is then imported and used in various `axios.get` and `axios.delete` calls in `components/meal-card.vue` and `pages/meal-page.vue` replacing the hardcoded URLs. This refactor is a step towards having a central location for the base API URL and improving maintainability.