Refactor #18: codes to improve readability and maintainability
Updated the code base to improve readability, and maintainability by moving API calls to a helper file. It reduces the repetition of the Axios calls and makes the code cleaner. This refactoring also helps when the base URL of the API changes, we only need to change it in one place.