docs #18: Add project documentation and UML schema
Created a new DEVELOPER.md file to provide guidelines for new developers on the project structure, flow of components, and dependencies. This documentation includes information on the purpose of directories, an overview of Vue.js components, workflow for meal-page, and more.

Also added a new UML.drawio.svg file under assets to provide a visual understanding of the project structure and interactions between components. Both files should aid new developers to understand and contribute to the project more effectively.