Tests: Fixing the test files to accommodate the merge
The urls in the test files for meals, food, and pets have been updated from "/Object" format to "/api/object". This was done to follow a more standard RESTful API design. The base_url variable in the test setups were also updated to reflect this change. In the test files, response.data is now accessed as response.data["results"] to handle paginated responses correctly. Some redundant parts of the code in test files have also been cleaned up.

Added the DateTimeField to the Meal Serializer, which is not required.